fix: :wrench: remove user path and normalize files
