+++ ResoniteMetricsCounter/Patch/Category.cs
@@ -1,6 +1,7 @@
-﻿namespace ResoniteMetricsCounter.Patch;
+namespace ResoniteMetricsCounter.Patch;
 internal static class Category
 {
     internal const string CORE = "resonite-profiler.core";
     internal const string PROFILER = "resonite-profiler.profiler";
 }
+
