+++ ResoniteMetricsCounter/ResoniteMetricsCounter.csproj
@@ -51,7 +51,6 @@
 	    <Pack>True</Pack>
 	    <PackagePath>\</PackagePath>
 	  </None>
-	  <None Include="C:\Users\esnya\Documents\rml_mods\ResoniteMetricsCounter\ResoniteMetricsCounter\.editorconfig" />
 	</ItemGroup>
 
 	<ItemGroup>
@@ -95,7 +94,4 @@
 		<Message Text="Attempting to copy $(TargetFileName) to $(ResonitePath)rml_mods\HotReloadMods" Importance="high" />
 		<Copy SourceFiles="$(TargetDir)$(TargetFileName)" DestinationFolder="$(ResonitePath)rml_mods\HotReloadMods" ContinueOnError="true" />
 	</Target>
-	<ItemGroup>
-	  <EditorConfigFiles Remove="C:\Users\esnya\Documents\rml_mods\ResoniteMetricsCounter\ResoniteMetricsCounter\.editorconfig" />
-	</ItemGroup>
 </Project>
